:package: Chore: add coverage collector
