+++ .github/workflows/ci.yml
@@ -13,6 +13,7 @@ jobs:
       - name: npm install, make test-coverage
         run: |
           npm install
+          npm run compile
           npm run test
 
       - name: Coveralls
