+++ package.json
@@ -305,6 +305,7 @@
     "@picgo/bump-version": "^1.0.3",
     "@types/glob": "^7.1.1",
     "@types/inquirer": "^6.5.0",
+    "@types/istanbul": "^0.4.30",
     "@types/mocha": "^7.0.1",
     "@types/node": "^12.11.7",
     "@types/request-promise-native": "^1.0.17",
@@ -313,14 +314,16 @@
     "@typescript-eslint/parser": "^2.18.0",
     "cz-conventional-changelog": "^3.1.0",
     "cz-customizable": "6.2.0",
+    "decache": "^4.5.1",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-plugin-prettier": "^3.1.2",
     "glob": "^7.1.6",
     "husky": "^4.2.3",
+    "istanbul": "^0.4.5",
     "mocha": "^7.0.1",
-    "nyc": "^15.0.0",
-    "prettier": "1.19.1",
+    "prettier": "^1.19.1",
+    "remap-istanbul": "^0.13.0",
     "typescript": "^3.7.5",
     "vscode-test": "^1.3.0"
   },
